requirements.txt
apache-airflow-providers-common-sql>=1.3.1
apache-airflow-providers-postgres>=5.3.1
apache-airflow-providers-slack
beautifulsoup4==4.11.1
requests>=2.28.0
yfinance==0.1.87



